CMake/macOs: Fix Translations installation
Currently when installing from source on macOs,
the translations cmake will always try to install them in
`/Contents/translations` (whatever the CMAKE_INSTALL_PREFIX may be set to)

This makes the installation process fail:

```
CMake Error at qCC/translations/cmake_install.cmake:49 (file):
  file cannot create directory: /Contents/translations.  Maybe need
  administrative privileges.
Call Stack (most recent call first):
  qCC/cmake_install.cmake:130 (include)
  cmake_install.cmake:44 (include)
```

This is because in the qCC\translations\CMakeLists.txt, the
destination path of translations is computed as
`${CLOUDCOMPARE_MAC_BASE_DIR}/Contents/translations`, however since
commit e2804cc55399eb871f9d350b28347babd0e04f48,
the CLOUDCOMPARE_MAC_BASE_DIR is no longer defined before it is used
in the translarions CMakeLists, so cmake treats it as an empty string.

The fix is simply to move the add_subdirectory(translation)
after the macOS specific variables are defined.